Refactor Project Utilities and Improve Repository Ownership Verification
New Utility File: project-helpers.ts
(Intentionally kept all the helper functions for future use)
Router Refactoring
early-submissions.ts: added enhanced user ownership checking
submissions.ts: added enhanced user ownership checking
projects.ts: Updated to use centralized utilities for ID resolution and status management
GitHub Driver: Improved ownership verification logic with better organization membership checking
GitLab Driver: Enhanced permission checking and access level validation
